add send mail

+++ src/database/migrations/1615891882418-CreateSurveysUsers.ts
@@ -21,7 +21,7 @@ export class CreateSurveysUsers1615891882418 implements MigrationInterface {
           },
           {
             name: 'value',
-            type: 'integer',
+            type: 'smallint',
             isNullable: true
           },
           {
